+++ src/main/java6/net/finmath/marketdata/model/curves/DiscountCurveFromProductOfCurves.java
@@ -6,9 +6,10 @@
 package net.finmath.marketdata.model.curves;
 
 import java.io.Serializable;
-import java.util.Calendar;
 
 import net.finmath.marketdata.model.AnalyticModelInterface;
+
+import org.joda.time.LocalDate;
 
 /**
  * A discount curve derived from other discount curves
@@ -29,7 +30,7 @@ public class DiscountCurveFromProductOfCurves extends AbstractCurve implements S
 	 * @param referenceDate The reference date of this curve.
      * @param curves Argument list or array of curves.
      */
-    public DiscountCurveFromProductOfCurves(String name, Calendar referenceDate, DiscountCurveInterface... curves) {
+    public DiscountCurveFromProductOfCurves(String name, LocalDate referenceDate, DiscountCurveInterface... curves) {
     	super(name, referenceDate);
 
     	this.curves = curves;