+++ src/main/java6/net/finmath/marketdata/model/volatilities/CapletVolatilities.java
@@ -12,7 +12,6 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.StringTokenizer;
@@ -24,6 +23,8 @@ import net.finmath.marketdata.model.curves.DiscountCurveInterface;
 import net.finmath.marketdata.model.curves.ForwardCurveInterface;
 import net.finmath.time.TimeDiscretization;
 import net.finmath.time.TimeDiscretizationInterface;
+
+import org.joda.time.LocalDate;
 
 /**
  * A very simple container for Caplet volatilities.
@@ -56,7 +57,7 @@ public class CapletVolatilities extends AbstractVolatilitySurface {
 	 * @param volatilityConvention The quoting convention of the volatilities provided.
 	 * @param discountCurve The associated discount curve.
 	 */
-	public CapletVolatilities(String name, Calendar referenceDate, ForwardCurveInterface forwardCurve,
+	public CapletVolatilities(String name, LocalDate referenceDate, ForwardCurveInterface forwardCurve,
 			double[] maturities,
 			double[] strikes,
 			double[] volatilities,
@@ -84,7 +85,7 @@ public class CapletVolatilities extends AbstractVolatilitySurface {
 	 * @param name The name of this volatility surface.
 	 * @param referenceDate The reference date for this volatility surface, i.e., the date which defined t=0.
 	 */
-	private CapletVolatilities(String name, Calendar referenceDate) {
+	private CapletVolatilities(String name, LocalDate referenceDate) {
 		super(name, referenceDate);
 	}