+++ src/main/java6/net/finmath/time/businessdaycalendar/BusinessdayCalendar.java
@@ -6,10 +6,10 @@
 
 package net.finmath.time.businessdaycalendar;
 
-import java.util.Calendar;
-import java.util.Date;
-import java.util.GregorianCalendar;
 import java.util.StringTokenizer;
+
+import org.joda.time.LocalDate;
+import org.joda.time.ReadablePeriod;
 
 /**
  * Base class for all business day calendars.
@@ -22,32 +22,32 @@ import java.util.StringTokenizer;
 public abstract class BusinessdayCalendar implements BusinessdayCalendarInterface {
 
 	/* (non-Javadoc)
-	 * @see net.finmath.time.BusinessdayCalendarInterface#getAdjustedDate(java.util.Calendar, net.finmath.time.BusinessdayCalendarInterface.DateRollConvention)
+	 * @see net.finmath.time.BusinessdayCalendarInterface#getAdjustedDate(java.time.LocalDate, net.finmath.time.BusinessdayCalendarInterface.DateRollConvention)
 	 */
 	@Override
-	public Calendar getAdjustedDate(Calendar date, DateRollConvention dateRollConvention) {
+	public LocalDate getAdjustedDate(LocalDate date, DateRollConvention dateRollConvention) {
 		if(dateRollConvention == DateRollConvention.UNADJUSTED) {
-			return (Calendar)date.clone();
+			return date;
 		}
 		else if(dateRollConvention == DateRollConvention.MODIFIED_FOLLOWING) {
-			Calendar adjustedDate = getAdjustedDate(date, DateRollConvention.FOLLOWING);
-			if(adjustedDate.get(Calendar.MONTH) != date.get(Calendar.MONTH)) {
+			LocalDate adjustedDate = getAdjustedDate(date, DateRollConvention.FOLLOWING);
+			if (adjustedDate.getMonthOfYear() != date.getMonthOfYear()){
 				return getAdjustedDate(date, DateRollConvention.PRECEDING);
 			}
 			else return adjustedDate;
 		}
 		else if(dateRollConvention == DateRollConvention.MODIFIED_PRECEDING) {
-			Calendar adjustedDate = getAdjustedDate(date, DateRollConvention.PRECEDING);
-			if(adjustedDate.get(Calendar.MONTH) != date.get(Calendar.MONTH)) {
+			LocalDate adjustedDate = getAdjustedDate(date, DateRollConvention.PRECEDING);
+			if (adjustedDate.getMonthOfYear() != date.getMonthOfYear()) {
 				return getAdjustedDate(date, DateRollConvention.FOLLOWING);
 			}
 			else return adjustedDate;
 		}
 		else if(dateRollConvention == DateRollConvention.FOLLOWING || dateRollConvention == DateRollConvention.PRECEDING) {
 			int adjustment = dateRollConvention == DateRollConvention.FOLLOWING ? 1 : -1;
-			Calendar adjustedDate = (Calendar)date.clone();
+			LocalDate adjustedDate = date;
 			while(!isBusinessday(adjustedDate)) {
-				adjustedDate.add(Calendar.DAY_OF_YEAR, adjustment);
+				adjustedDate = adjustedDate.plusDays(adjustment);
 			}
 			return adjustedDate;
 		}
@@ -55,22 +55,16 @@ public abstract class BusinessdayCalendar implements BusinessdayCalendarInterfac
 		throw new IllegalArgumentException("Unknown date roll convention.");
 	}
 	
-	/* (non-Javadoc)
-	 * @see net.finmath.time.businessdaycalendar.BusinessdayCalendarInterface#getAdjustedDate(java.util.Calendar, java.lang.String, net.finmath.time.businessdaycalendar.BusinessdayCalendarInterface.DateRollConvention)
-	 */
-	public Calendar getAdjustedDate(Calendar baseDate, String dateOffsetCode, DateRollConvention dateRollConvention) {
-		return this.getAdjustedDate(createDateFromDateAndOffsetCode(baseDate, dateOffsetCode), dateRollConvention);
-	}
 	
 	/* (non-Javadoc)
-	 * @see net.finmath.time.businessdaycalendar.BusinessdayCalendarInterface#getRolledDate(java.util.Calendar, int)
+	 * @see net.finmath.time.businessdaycalendar.BusinessdayCalendarInterface#getRolledDate(java.time.LocalDate, int)
 	 */
-	public Calendar getRolledDate(Calendar baseDate, int businessDays) {
-		Calendar			rolledDate			= (Calendar)baseDate.clone();
+	public LocalDate getRolledDate(LocalDate baseDate, int businessDays) {
+		LocalDate			rolledDate			= baseDate;
 		int					direction			= businessDays >= 0 ? 1: -1;
 		DateRollConvention	dateRollConvention	= direction > 0 ? DateRollConvention.FOLLOWING : DateRollConvention.PRECEDING;
 		while(businessDays != 0) {
-			rolledDate.add(Calendar.DAY_OF_YEAR, direction);
+			rolledDate = rolledDate.plusDays(direction);
 			rolledDate = getAdjustedDate(rolledDate, dateRollConvention);
 			businessDays -= direction;
 		}
@@ -92,10 +86,8 @@ public abstract class BusinessdayCalendar implements BusinessdayCalendarInterfac
 	 * @param dateRollConvention The date roll convention to be used for the adjustment.
 	 * @return The adjusted date applying dateRollConvention to the given date.
 	 */
-	public Date getAdjustedDate(Date baseDate, String dateOffsetCode, DateRollConvention dateRollConvention) {
-		Calendar baseDateAsCalendar = GregorianCalendar.getInstance();
-		baseDateAsCalendar.setTime(baseDate);
-		return this.getAdjustedDate(createDateFromDateAndOffsetCode(baseDateAsCalendar, dateOffsetCode), dateRollConvention).getTime();
+	public LocalDate getAdjustedDate(LocalDate baseDate, String dateOffsetCode, DateRollConvention dateRollConvention) {
+		return this.getAdjustedDate(createDateFromDateAndOffsetCode(baseDate, dateOffsetCode), dateRollConvention);
 	}
 
 	/**
@@ -116,42 +108,46 @@ public abstract class BusinessdayCalendar implements BusinessdayCalendarInterfac
 	 * @param dateOffsetCode String containing date offset codes (like 2D, 1W, 3M, etc.) or combination of them separated by spaces.
 	 * @return A date corresponding the date adding the offset to the start date.
 	 */
-	public static Calendar createDateFromDateAndOffsetCode(Calendar baseDate, String dateOffsetCode) {
+	public static LocalDate createDateFromDateAndOffsetCode(LocalDate baseDate, String dateOffsetCode) {
 		dateOffsetCode = dateOffsetCode.trim();
 
 		StringTokenizer tokenizer = new StringTokenizer(dateOffsetCode);
 		
-		Calendar maturity = (Calendar)baseDate.clone();
+		LocalDate maturity = baseDate;
 		while(tokenizer.hasMoreTokens()) {
 			String maturityCodeSingle = tokenizer.nextToken();
 
-			char	unitChar		= maturityCodeSingle.toLowerCase().charAt(maturityCodeSingle.length()-1);
-			int		unit;
+			char unitChar = maturityCodeSingle.toLowerCase().charAt(maturityCodeSingle.length()-1);
+
 			switch(unitChar) {
 			case 'd':
-				unit = Calendar.DAY_OF_YEAR;
+			{
+				int maturityValue = Integer.valueOf(maturityCodeSingle.substring(0, maturityCodeSingle.length()-1));
+				maturity = maturity.plusDays(maturityValue);
 				break;
+			}
 			case 'w':
-				unit = Calendar.WEEK_OF_YEAR;
+			{
+				int maturityValue = Integer.valueOf(maturityCodeSingle.substring(0, maturityCodeSingle.length()-1));
+				maturity = maturity.plusWeeks(maturityValue);
 				break;
+			}
 			case 'm':
-				unit = Calendar.MONTH;
+			{
+				int maturityValue = Integer.valueOf(maturityCodeSingle.substring(0, maturityCodeSingle.length()-1));
+				maturity = maturity.plusMonths(maturityValue);
 				break;
+			}
 			case 'y':
-				unit = Calendar.YEAR;
+			{
+				int maturityValue = Integer.valueOf(maturityCodeSingle.substring(0, maturityCodeSingle.length()-1));
+				maturity = maturity.plusYears(maturityValue);
 				break;
+			}
 			default:
-				unit = Calendar.FIELD_COUNT;
-			}
-
-			if(unit != Calendar.FIELD_COUNT) {
-				int maturityValue		= Integer.valueOf(maturityCodeSingle.substring(0, maturityCodeSingle.length()-1));
-				maturity.add(unit, maturityValue);
-			}
-			else {
 				// Try to parse a double as ACT/365
 				double maturityValue	= Double.valueOf(maturityCodeSingle);
-				maturity.add(Calendar.DAY_OF_YEAR, (int)Math.round(maturityValue * 365));
+				maturity = maturity.plusDays((int)Math.round(maturityValue * 365));
 			}
 		}