+++ src/main/java6/net/finmath/time/daycount/DayCountConvention_ACT_365L.java
@@ -6,16 +6,14 @@
 
 package net.finmath.time.daycount;
 
-import java.util.Calendar;
-import java.util.GregorianCalendar;
+import org.joda.time.LocalDate;
 
 /**
  * Implementation of ACT/365L.
  * 
  * Calculates the day count by calculating the actual number of days between startDate and endDate.
  * 
- * The method is only exact, if the two calendar dates are (approximately) on the same time. A fractional day is
- * rounded to the approximately nearest day (since daylight saving is not considered, the notion of nearest may be off by one hour).
+ * A fractional day is rounded to the approximately nearest day.
  * 
  * The day count fraction is calculated using ACT/365L convention, that is, the
  * day count is divided by 365 if the end date is not a leap year and 366 if the end date is a leap year.
@@ -34,14 +32,13 @@ public class DayCountConvention_ACT_365L extends DayCountConvention_ACT {
 	}
 	
 	@Override
-	public double getDaycountFraction(Calendar startDate, Calendar endDate) {
-		if(startDate.after(endDate)) return -getDaycountFraction(endDate,startDate);
+	public double getDaycountFraction(LocalDate startDate, LocalDate endDate) {
+		if(startDate.isAfter(endDate)) return -getDaycountFraction(endDate,startDate);
 
 		double daysPerYear = 365.0;
 		
 		// Check endDate for leap year
-		GregorianCalendar gregorianCalendar = new GregorianCalendar();
-		if(gregorianCalendar.isLeapYear(endDate.get(Calendar.YEAR))) daysPerYear = 366.0;
+		if (endDate.year().isLeap()) daysPerYear = 366.0;
 		
 		double daycountFraction = getDaycount(startDate, endDate) / daysPerYear;