+++ src/test/java6/net/finmath/montecarlo/interestrate/products/components/ExposureTest.java
@@ -9,7 +9,6 @@ package net.finmath.montecarlo.interestrate.products.components;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.text.NumberFormat;
-import java.util.GregorianCalendar;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
@@ -41,6 +40,8 @@ import net.finmath.time.TimeDiscretization;
 import net.finmath.time.businessdaycalendar.BusinessdayCalendarExcludingTARGETHolidays;
 import net.finmath.time.businessdaycalendar.BusinessdayCalendarInterface;
 
+import org.joda.time.DateTimeConstants;
+import org.joda.time.LocalDate;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -58,9 +59,9 @@ public class ExposureTest {
 		 * Create a receiver swap (receive fix, pay float)
 		 */
 		ScheduleInterface legScheduleRec = ScheduleGenerator.createScheduleFromConventions(
-				new GregorianCalendar(2015,GregorianCalendar.JANUARY,03) /* referenceDate */,
-				new GregorianCalendar(2015,GregorianCalendar.JANUARY,06) /* startDate */,
-				new GregorianCalendar(2025,GregorianCalendar.JANUARY,06) /* maturityDate */,
+				new LocalDate(2015, DateTimeConstants.JANUARY, 03) /* referenceDate */, 
+				new LocalDate(2015, DateTimeConstants.JANUARY, 06) /* startDate */,
+				new LocalDate(2015, DateTimeConstants.JANUARY, 06) /* maturityDate */,
 				ScheduleGenerator.Frequency.ANNUAL /* frequency */,
 				ScheduleGenerator.DaycountConvention.ACT_365 /* daycountConvention */,
 				ScheduleGenerator.ShortPeriodConvention.FIRST /* shortPeriodConvention */,
@@ -70,9 +71,9 @@ public class ExposureTest {
 				0 /* paymentOffsetDays */);
 
 		ScheduleInterface legSchedulePay = ScheduleGenerator.createScheduleFromConventions(
-				new GregorianCalendar(2015,GregorianCalendar.JANUARY,03) /* referenceDate */,
-				new GregorianCalendar(2015,GregorianCalendar.JANUARY,06) /* startDate */,
-				new GregorianCalendar(2025,GregorianCalendar.JANUARY,06) /* maturityDate */,
+				new LocalDate(2015, DateTimeConstants.JANUARY, 03) /* referenceDate */, 
+				new LocalDate(2015, DateTimeConstants.JANUARY, 06) /* startDate */,
+				new LocalDate(2015, DateTimeConstants.JANUARY, 06) /* maturityDate */,
 				ScheduleGenerator.Frequency.QUARTERLY /* frequency */,
 				ScheduleGenerator.DaycountConvention.ACT_365 /* daycountConvention */,
 				ScheduleGenerator.ShortPeriodConvention.FIRST /* shortPeriodConvention */,